Moved some files and code around.  No functional changes.
