+++ Unit Tests/OKP/TestUtilities/Doubles/JoinedRoomsProviderDelegate.swift
