+++ Unit Tests/OKP/TestUtilities/Doubles/SDKInfoProvider.swift
