+++ Unit Tests/OKP/TestUtilities/Doubles/StoreDelegate.swift
