+++ Unit Tests/OKP/TestUtilities/Doubles/StoreListener.swift
