+++ Unit Tests/OKP/TestUtilities/Doubles/SubscriptionDelegate.swift
